[proliferate] split session store modules
