--- anyharness/crates/anyharness-lib/src/sessions/store.rs
@@ -1,2836 +1,0 @@
-use rusqlite::{params, OptionalExtension};
-
-use super::model::{
-    PendingConfigChangeRecord, PendingPromptRecord, PromptAttachmentKind, PromptAttachmentRecord,
-    PromptAttachmentSource, PromptAttachmentState, SessionBackgroundWorkRecord,
-    SessionBackgroundWorkState, SessionBackgroundWorkTrackerKind, SessionEventRecord,
-    SessionLiveConfigSnapshotRecord, SessionMcpBindingPolicy, SessionRawNotificationRecord,
-    SessionRecord,
-};
-use crate::acp::persistence_sanitizer::{
-    sanitize_raw_notification_json_for_sqlite, sanitize_session_event_for_sqlite,
-};
-use crate::origin::{decode_origin_json, encode_origin_json};
-use crate::persistence::Db;
-use crate::sessions::links::model::{SessionLinkRecord, SessionLinkRelation};
-use crate::sessions::prompt::PromptPayload;
-
-#[derive(Clone)]
-pub struct SessionStore {
-    db: Db,
-}
-
-impl SessionStore {
-    pub fn new(db: Db) -> Self {
-        Self { db }
-    }
-
-    pub fn insert(&self, record: &SessionRecord) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            insert_session_row(conn, record)?;
-            Ok(())
-        })
-    }
-
-    pub fn insert_session_with_link(
-        &self,
-        record: &SessionRecord,
-        link: &SessionLinkRecord,
-    ) -> anyhow::Result<()> {
-        self.db.with_tx(|conn| {
-            insert_session_row(conn, record)?;
-            insert_session_link_row(conn, link)?;
-            Ok(())
-        })
-    }
-
-    pub fn insert_fork_session_with_link_and_event_snapshot(
-        &self,
-        record: &SessionRecord,
-        link: &SessionLinkRecord,
-    ) -> anyhow::Result<usize> {
-        anyhow::ensure!(
-            link.relation == SessionLinkRelation::Fork,
-            "event snapshots are only supported for fork links"
-        );
-        anyhow::ensure!(
-            link.child_session_id == record.id,
-            "fork link child id must match inserted session"
-        );
-
-        self.db.with_tx(|conn| {
-            insert_session_row(conn, record)?;
-            insert_session_link_row(conn, link)?;
-            let copied = conn.execute(
-                "INSERT INTO session_events (
-                    session_id, seq, timestamp, event_type, turn_id, item_id, payload_json
-                 )
-                 SELECT ?1, seq, timestamp, event_type, turn_id, item_id, payload_json
-                 FROM session_events
-                 WHERE session_id = ?2
-                 ORDER BY seq ASC",
-                params![record.id, link.parent_session_id],
-            )?;
-            Ok(copied)
-        })
-    }
-
-    pub fn delete_session(&self, id: &str) -> anyhow::Result<()> {
-        self.db.with_tx(|conn| delete_session_in_tx(conn, id))
-    }
-
-    pub fn find_by_id(&self, id: &str) -> anyhow::Result<Option<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row("SELECT * FROM sessions WHERE id = ?1", [id], |row| {
-                map_session(row)
-            })
-            .optional()
-        })
-    }
-
-    pub fn update_mcp_bindings(
-        &self,
-        id: &str,
-        mcp_bindings_ciphertext: Option<String>,
-        mcp_binding_summaries_json: Option<String>,
-    ) -> anyhow::Result<()> {
-        let now = chrono::Utc::now().to_rfc3339();
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions
-                 SET mcp_bindings_ciphertext = ?1,
-                     mcp_binding_summaries_json = ?2,
-                     updated_at = ?3
-                 WHERE id = ?4",
-                params![mcp_bindings_ciphertext, mcp_binding_summaries_json, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_mcp_binding_summaries(
-        &self,
-        id: &str,
-        mcp_binding_summaries_json: Option<String>,
-    ) -> anyhow::Result<()> {
-        let now = chrono::Utc::now().to_rfc3339();
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions
-                 SET mcp_binding_summaries_json = ?1,
-                     updated_at = ?2
-                 WHERE id = ?3",
-                params![mcp_binding_summaries_json, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_action_capabilities_json(
-        &self,
-        id: &str,
-        action_capabilities_json: Option<String>,
-        now: &str,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions
-                 SET action_capabilities_json = ?1,
-                     updated_at = ?2
-                 WHERE id = ?3",
-                params![action_capabilities_json, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn has_inbound_link_relation(
-        &self,
-        session_id: &str,
-        relation: SessionLinkRelation,
-    ) -> anyhow::Result<bool> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT EXISTS(
-                    SELECT 1 FROM session_links
-                    WHERE child_session_id = ?1 AND relation = ?2
-                )",
-                params![session_id, relation.as_str()],
-                |row| row.get(0),
-            )
-            .map_err(Into::into)
-        })
-    }
-
-    pub fn find_parent_by_inbound_link_relation(
-        &self,
-        session_id: &str,
-        relation: SessionLinkRelation,
-    ) -> anyhow::Result<Option<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT sessions.*
-                 FROM session_links
-                 JOIN sessions ON sessions.id = session_links.parent_session_id
-                 WHERE session_links.child_session_id = ?1
-                   AND session_links.relation = ?2
-                 ORDER BY session_links.created_at ASC, session_links.id ASC
-                 LIMIT 1",
-                params![session_id, relation.as_str()],
-                map_session,
-            )
-            .optional()
-            .map_err(Into::into)
-        })
-    }
-
-    pub fn list_by_workspace(&self, workspace_id: &str) -> anyhow::Result<Vec<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM sessions WHERE workspace_id = ?1 ORDER BY updated_at DESC",
-            )?;
-            let rows = stmt.query_map([workspace_id], |row| map_session(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_all(&self) -> anyhow::Result<Vec<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare("SELECT * FROM sessions ORDER BY updated_at DESC")?;
-            let rows = stmt.query_map([], |row| map_session(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_visible_by_workspace(
-        &self,
-        workspace_id: &str,
-    ) -> anyhow::Result<Vec<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM sessions
-                 WHERE workspace_id = ?1 AND dismissed_at IS NULL AND closed_at IS NULL
-                 ORDER BY updated_at DESC",
-            )?;
-            let rows = stmt.query_map([workspace_id], map_session)?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_with_dismissed_by_workspace(
-        &self,
-        workspace_id: &str,
-    ) -> anyhow::Result<Vec<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM sessions
-                 WHERE workspace_id = ?1 AND closed_at IS NULL
-                 ORDER BY updated_at DESC",
-            )?;
-            let rows = stmt.query_map([workspace_id], map_session)?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_visible_all(&self) -> anyhow::Result<Vec<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM sessions
-                 WHERE dismissed_at IS NULL AND closed_at IS NULL
-                 ORDER BY updated_at DESC",
-            )?;
-            let rows = stmt.query_map([], map_session)?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_with_dismissed_all(&self) -> anyhow::Result<Vec<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM sessions
-                 WHERE closed_at IS NULL
-                 ORDER BY updated_at DESC",
-            )?;
-            let rows = stmt.query_map([], map_session)?;
-            rows.collect()
-        })
-    }
-
-    pub fn update_status(&self, id: &str, status: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET status = ?1, updated_at = ?2 WHERE id = ?3",
-                params![status, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_native_session_id(
-        &self,
-        id: &str,
-        native_session_id: &str,
-        now: &str,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET native_session_id = ?1, updated_at = ?2 WHERE id = ?3",
-                params![native_session_id, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn clear_native_session_id(&self, id: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET native_session_id = NULL, updated_at = ?1 WHERE id = ?2",
-                params![now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_last_prompt_at(&self, id: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET last_prompt_at = ?1, updated_at = ?1 WHERE id = ?2",
-                params![now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_requested_configuration(
-        &self,
-        id: &str,
-        requested_model_id: Option<&str>,
-        requested_mode_id: Option<&str>,
-        now: &str,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions
-                 SET requested_model_id = ?1, requested_mode_id = ?2, updated_at = ?3
-                 WHERE id = ?4",
-                params![requested_model_id, requested_mode_id, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_current_configuration(
-        &self,
-        id: &str,
-        current_model_id: Option<&str>,
-        current_mode_id: Option<&str>,
-        now: &str,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions
-                 SET current_model_id = ?1, current_mode_id = ?2, updated_at = ?3
-                 WHERE id = ?4",
-                params![current_model_id, current_mode_id, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn update_title(&self, id: &str, title: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET title = ?1, updated_at = ?2 WHERE id = ?3",
-                params![title, now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn mark_closed(&self, id: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET status = 'closed', closed_at = ?1, updated_at = ?1 WHERE id = ?2",
-                params![now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn mark_dismissed(&self, id: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions
-                 SET dismissed_at = COALESCE(dismissed_at, ?1),
-                     updated_at = CASE WHEN dismissed_at IS NULL THEN ?1 ELSE updated_at END
-                 WHERE id = ?2",
-                params![now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn clear_dismissed(&self, id: &str, now: &str) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE sessions SET dismissed_at = NULL, updated_at = ?1 WHERE id = ?2",
-                params![now, id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn pop_last_dismissed_in_workspace(
-        &self,
-        workspace_id: &str,
-        now: &str,
-    ) -> anyhow::Result<Option<SessionRecord>> {
-        self.db.with_tx(|conn| {
-            let record = conn
-                .query_row(
-                    "SELECT * FROM sessions
-                     WHERE workspace_id = ?1 AND dismissed_at IS NOT NULL AND closed_at IS NULL
-                     ORDER BY dismissed_at DESC, updated_at DESC
-                     LIMIT 1",
-                    [workspace_id],
-                    map_session,
-                )
-                .optional()?;
-
-            let Some(record) = record else {
-                return Ok(None);
-            };
-
-            conn.execute(
-                "UPDATE sessions SET dismissed_at = NULL, updated_at = ?1 WHERE id = ?2",
-                params![now, record.id],
-            )?;
-
-            let restored = conn
-                .query_row(
-                    "SELECT * FROM sessions WHERE id = ?1",
-                    [&record.id],
-                    map_session,
-                )
-                .optional()?;
-            Ok(restored)
-        })
-    }
-
-    pub fn find_last_dismissed_in_workspace(
-        &self,
-        workspace_id: &str,
-    ) -> anyhow::Result<Option<SessionRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT * FROM sessions
-                 WHERE workspace_id = ?1 AND dismissed_at IS NOT NULL AND closed_at IS NULL
-                 ORDER BY dismissed_at DESC, updated_at DESC
-                 LIMIT 1",
-                [workspace_id],
-                map_session,
-            )
-            .optional()
-        })
-    }
-
-    pub fn upsert_live_config_snapshot(
-        &self,
-        record: &SessionLiveConfigSnapshotRecord,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO session_live_config_snapshots (
-                    session_id, source_seq, raw_config_options_json, normalized_controls_json,
-                    prompt_capabilities_json, updated_at
-                 ) VALUES (?1, ?2, ?3, ?4, ?5, ?6)
-                 ON CONFLICT(session_id) DO UPDATE SET
-                    source_seq = excluded.source_seq,
-                    raw_config_options_json = excluded.raw_config_options_json,
-                    normalized_controls_json = excluded.normalized_controls_json,
-                    prompt_capabilities_json = excluded.prompt_capabilities_json,
-                    updated_at = excluded.updated_at",
-                params![
-                    record.session_id,
-                    record.source_seq,
-                    record.raw_config_options_json,
-                    record.normalized_controls_json,
-                    record.prompt_capabilities_json,
-                    record.updated_at,
-                ],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn import_bundle(
-        &self,
-        session: &SessionRecord,
-        live_config_snapshot: Option<&SessionLiveConfigSnapshotRecord>,
-        pending_config_changes: &[PendingConfigChangeRecord],
-        pending_prompts: &[PendingPromptRecord],
-        prompt_attachments: &[PromptAttachmentRecord],
-        events: &[SessionEventRecord],
-        raw_notifications: &[SessionRawNotificationRecord],
-    ) -> anyhow::Result<()> {
-        self.db.with_tx(|conn| {
-            insert_session_row(conn, session)?;
-            if let Some(snapshot) = live_config_snapshot {
-                upsert_live_config_snapshot_row(conn, snapshot)?;
-            }
-            for change in pending_config_changes {
-                upsert_pending_config_change_row(conn, change)?;
-            }
-            for prompt in pending_prompts {
-                insert_pending_prompt_row(conn, prompt)?;
-            }
-            for attachment in prompt_attachments {
-                insert_prompt_attachment_row(conn, attachment)?;
-            }
-            for event in events {
-                insert_event_row(conn, event)?;
-            }
-            for notification in raw_notifications {
-                insert_raw_notification_row(conn, notification)?;
-            }
-            Ok(())
-        })
-    }
-
-    pub fn find_live_config_snapshot(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Option<SessionLiveConfigSnapshotRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT * FROM session_live_config_snapshots WHERE session_id = ?1",
-                [session_id],
-                map_live_config_snapshot,
-            )
-            .optional()
-        })
-    }
-
-    pub fn upsert_pending_config_change(
-        &self,
-        record: &PendingConfigChangeRecord,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO session_pending_config_changes (session_id, config_id, value, queued_at)
-                 VALUES (?1, ?2, ?3, ?4)
-                 ON CONFLICT(session_id, config_id) DO UPDATE SET
-                    value = excluded.value,
-                    queued_at = excluded.queued_at",
-                params![record.session_id, record.config_id, record.value, record.queued_at],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn list_pending_config_changes(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Vec<PendingConfigChangeRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_pending_config_changes
-                 WHERE session_id = ?1
-                 ORDER BY queued_at ASC, config_id ASC",
-            )?;
-            let rows = stmt.query_map([session_id], map_pending_config_change)?;
-            rows.collect()
-        })
-    }
-
-    pub fn delete_pending_config_change(
-        &self,
-        session_id: &str,
-        config_id: &str,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "DELETE FROM session_pending_config_changes WHERE session_id = ?1 AND config_id = ?2",
-                params![session_id, config_id],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn insert_pending_prompt(
-        &self,
-        session_id: &str,
-        text: &str,
-        prompt_id: Option<&str>,
-    ) -> anyhow::Result<PendingPromptRecord> {
-        self.insert_pending_prompt_payload(
-            session_id,
-            &PromptPayload::text(text.to_string()),
-            prompt_id,
-        )
-    }
-
-    pub fn insert_pending_prompt_payload(
-        &self,
-        session_id: &str,
-        payload: &PromptPayload,
-        prompt_id: Option<&str>,
-    ) -> anyhow::Result<PendingPromptRecord> {
-        let queued_at = chrono::Utc::now().to_rfc3339();
-        let blocks_json = payload.blocks_json()?;
-        let provenance_json = payload.provenance_json()?;
-        self.db.with_tx(|tx| {
-            let next_seq: i64 = tx.query_row(
-                "SELECT COALESCE(MAX(seq), 0) + 1 FROM session_pending_prompts WHERE session_id = ?1",
-                [session_id],
-                |row| row.get(0),
-            )?;
-            tx.execute(
-                "INSERT INTO session_pending_prompts (
-                    session_id, seq, prompt_id, text, blocks_json, provenance_json, queued_at
-                 ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-                params![
-                    session_id,
-                    next_seq,
-                    prompt_id,
-                    payload.text_summary,
-                    blocks_json,
-                    provenance_json,
-                    queued_at
-                ],
-            )?;
-            Ok(PendingPromptRecord {
-                session_id: session_id.to_string(),
-                seq: next_seq,
-                prompt_id: prompt_id.map(|s| s.to_string()),
-                text: payload.text_summary.clone(),
-                blocks_json,
-                provenance_json,
-                queued_at: queued_at.clone(),
-            })
-        })
-    }
-
-    pub fn list_pending_prompts(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Vec<PendingPromptRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_pending_prompts
-                 WHERE session_id = ?1
-                 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map([session_id], map_pending_prompt)?;
-            rows.collect()
-        })
-    }
-
-    pub fn peek_head_pending_prompt(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Option<PendingPromptRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT * FROM session_pending_prompts
-                 WHERE session_id = ?1
-                 ORDER BY seq ASC
-                 LIMIT 1",
-                [session_id],
-                map_pending_prompt,
-            )
-            .optional()
-        })
-    }
-
-    pub fn find_pending_prompt(
-        &self,
-        session_id: &str,
-        seq: i64,
-    ) -> anyhow::Result<Option<PendingPromptRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT * FROM session_pending_prompts
-                 WHERE session_id = ?1 AND seq = ?2",
-                params![session_id, seq],
-                map_pending_prompt,
-            )
-            .optional()
-        })
-    }
-
-    pub fn update_pending_prompt_text(
-        &self,
-        session_id: &str,
-        seq: i64,
-        text: &str,
-    ) -> anyhow::Result<bool> {
-        self.update_pending_prompt_payload(session_id, seq, &PromptPayload::text(text.to_string()))
-    }
-
-    pub fn update_pending_prompt_payload(
-        &self,
-        session_id: &str,
-        seq: i64,
-        payload: &PromptPayload,
-    ) -> anyhow::Result<bool> {
-        let blocks_json = payload.blocks_json()?;
-        self.db.with_conn(|conn| {
-            let rows = conn.execute(
-                "UPDATE session_pending_prompts
-                 SET text = ?3, blocks_json = ?4
-                 WHERE session_id = ?1 AND seq = ?2",
-                params![session_id, seq, payload.text_summary, blocks_json],
-            )?;
-            Ok(rows > 0)
-        })
-    }
-
-    pub fn delete_pending_prompt(&self, session_id: &str, seq: i64) -> anyhow::Result<bool> {
-        Ok(self
-            .delete_pending_prompt_record(session_id, seq)?
-            .is_some())
-    }
-
-    pub fn delete_pending_prompt_record(
-        &self,
-        session_id: &str,
-        seq: i64,
-    ) -> anyhow::Result<Option<PendingPromptRecord>> {
-        self.db.with_conn(|conn| {
-            let record = conn
-                .query_row(
-                    "SELECT * FROM session_pending_prompts WHERE session_id = ?1 AND seq = ?2",
-                    params![session_id, seq],
-                    map_pending_prompt,
-                )
-                .optional()?;
-            if record.is_none() {
-                return Ok(None);
-            }
-            conn.execute(
-                "DELETE FROM session_pending_prompts WHERE session_id = ?1 AND seq = ?2",
-                params![session_id, seq],
-            )?;
-            Ok(record)
-        })
-    }
-
-    pub fn insert_prompt_attachment(&self, record: &PromptAttachmentRecord) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            insert_prompt_attachment_row(conn, record)?;
-            Ok(())
-        })
-    }
-
-    pub fn find_prompt_attachment(
-        &self,
-        session_id: &str,
-        attachment_id: &str,
-    ) -> anyhow::Result<Option<PromptAttachmentRecord>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT * FROM session_prompt_attachments
-                 WHERE session_id = ?1 AND attachment_id = ?2",
-                params![session_id, attachment_id],
-                map_prompt_attachment,
-            )
-            .optional()
-        })
-    }
-
-    pub fn list_prompt_attachments(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Vec<PromptAttachmentRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_prompt_attachments
-                 WHERE session_id = ?1
-                 ORDER BY created_at ASC, attachment_id ASC",
-            )?;
-            let rows = stmt.query_map([session_id], map_prompt_attachment)?;
-            rows.collect()
-        })
-    }
-
-    pub fn read_legacy_prompt_attachment_content(
-        &self,
-        session_id: &str,
-        attachment_id: &str,
-    ) -> anyhow::Result<Option<Vec<u8>>> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT content FROM session_prompt_attachments
-                 WHERE session_id = ?1 AND attachment_id = ?2",
-                params![session_id, attachment_id],
-                |row| row.get(0),
-            )
-            .optional()
-        })
-    }
-
-    pub fn update_prompt_attachment_storage_path(
-        &self,
-        session_id: &str,
-        attachment_id: &str,
-        storage_path: &str,
-    ) -> anyhow::Result<()> {
-        let now = chrono::Utc::now().to_rfc3339();
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE session_prompt_attachments
-                 SET storage_path = ?3, updated_at = ?4
-                 WHERE session_id = ?1 AND attachment_id = ?2",
-                params![session_id, attachment_id, storage_path, now],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn mark_prompt_attachments_state(
-        &self,
-        session_id: &str,
-        attachment_ids: &[String],
-        state: PromptAttachmentState,
-    ) -> anyhow::Result<()> {
-        if attachment_ids.is_empty() {
-            return Ok(());
-        }
-        let now = chrono::Utc::now().to_rfc3339();
-        self.db.with_tx(|conn| {
-            for attachment_id in attachment_ids {
-                conn.execute(
-                    "UPDATE session_prompt_attachments
-                     SET state = ?3, updated_at = ?4
-                     WHERE session_id = ?1 AND attachment_id = ?2",
-                    params![session_id, attachment_id, state.as_str(), now],
-                )?;
-            }
-            Ok(())
-        })
-    }
-
-    pub fn delete_prompt_attachments(
-        &self,
-        session_id: &str,
-        attachment_ids: &[&str],
-    ) -> anyhow::Result<()> {
-        if attachment_ids.is_empty() {
-            return Ok(());
-        }
-        self.db.with_tx(|conn| {
-            for attachment_id in attachment_ids {
-                conn.execute(
-                    "DELETE FROM session_prompt_attachments
-                     WHERE session_id = ?1 AND attachment_id = ?2 AND state = 'pending'",
-                    params![session_id, attachment_id],
-                )?;
-            }
-            Ok(())
-        })
-    }
-
-    pub fn upsert_or_refresh_pending_background_work(
-        &self,
-        record: &SessionBackgroundWorkRecord,
-    ) -> anyhow::Result<bool> {
-        self.db.with_conn(|conn| {
-            let rows = conn.execute(
-                "INSERT INTO session_background_work (
-                    session_id, tool_call_id, turn_id, tracker_kind, source_agent_kind, agent_id,
-                    output_file, state, created_at, updated_at, launched_at, last_activity_at, completed_at
-                 ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13)
-                 ON CONFLICT(session_id, tool_call_id) DO UPDATE SET
-                    turn_id = excluded.turn_id,
-                    tracker_kind = excluded.tracker_kind,
-                    source_agent_kind = excluded.source_agent_kind,
-                    agent_id = excluded.agent_id,
-                    output_file = excluded.output_file,
-                    updated_at = excluded.updated_at,
-                    launched_at = excluded.launched_at,
-                    last_activity_at = excluded.last_activity_at,
-                    completed_at = excluded.completed_at
-                 WHERE session_background_work.state = 'pending'",
-                params![
-                    record.session_id,
-                    record.tool_call_id,
-                    record.turn_id,
-                    record.tracker_kind.as_str(),
-                    record.source_agent_kind,
-                    record.agent_id,
-                    record.output_file,
-                    record.state.as_str(),
-                    record.created_at,
-                    record.updated_at,
-                    record.launched_at,
-                    record.last_activity_at,
-                    record.completed_at,
-                ],
-            )?;
-            Ok(rows > 0)
-        })
-    }
-
-    pub fn list_pending_background_work(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Vec<SessionBackgroundWorkRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_background_work
-                 WHERE session_id = ?1 AND state = 'pending'
-                 ORDER BY launched_at ASC, tool_call_id ASC",
-            )?;
-            let rows = stmt.query_map([session_id], map_background_work)?;
-            rows.collect()
-        })
-    }
-
-    pub fn touch_background_work_activity(
-        &self,
-        session_id: &str,
-        tool_call_id: &str,
-        last_activity_at: &str,
-    ) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "UPDATE session_background_work
-                 SET last_activity_at = ?3,
-                     updated_at = ?3
-                 WHERE session_id = ?1 AND tool_call_id = ?2 AND state = 'pending'",
-                params![session_id, tool_call_id, last_activity_at],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn mark_background_work_terminal(
-        &self,
-        session_id: &str,
-        tool_call_id: &str,
-        state: SessionBackgroundWorkState,
-        completed_at: &str,
-    ) -> anyhow::Result<bool> {
-        self.db.with_conn(|conn| {
-            let rows = conn.execute(
-                "UPDATE session_background_work
-                 SET state = ?3,
-                     updated_at = ?4,
-                     completed_at = ?4
-                 WHERE session_id = ?1 AND tool_call_id = ?2 AND state = 'pending'",
-                params![session_id, tool_call_id, state.as_str(), completed_at],
-            )?;
-            Ok(rows > 0)
-        })
-    }
-
-    pub fn next_event_seq(&self, session_id: &str) -> anyhow::Result<i64> {
-        self.db.with_conn(|conn| {
-            let max: Option<i64> = conn.query_row(
-                "SELECT MAX(seq) FROM session_events WHERE session_id = ?1",
-                [session_id],
-                |row| row.get(0),
-            )?;
-            Ok(max.unwrap_or(0) + 1)
-        })
-    }
-
-    pub fn append_event(&self, event: &SessionEventRecord) -> anyhow::Result<()> {
-        self.db.with_conn(|conn| {
-            insert_event_row(conn, event)?;
-            Ok(())
-        })
-    }
-
-    pub(crate) fn append_event_and_touch_session(
-        &self,
-        event: &SessionEventRecord,
-    ) -> anyhow::Result<()> {
-        self.db.with_tx(|conn| {
-            insert_event_row(conn, event)?;
-            conn.execute(
-                "UPDATE sessions SET updated_at = ?1 WHERE id = ?2",
-                params![event.timestamp, event.session_id],
-            )?;
-            Ok(())
-        })
-    }
-
-    /// Atomically append an already-normalized runtime-owned event using the
-    /// next durable sequence number.
-    ///
-    /// # Invariants
-    ///
-    /// Callers must hold the ACP start/inject critical section described in
-    /// `docs/anyharness/src/acp.md#startinject-sequence-invariant` and must
-    /// have confirmed that no live actor owns event sequencing for this
-    /// session. Live actors assign seq from memory, so calling this while an
-    /// actor is live can race that actor's `SessionEventSink`.
-    pub(crate) fn append_event_with_next_seq(
-        &self,
-        session_id: &str,
-        event: anyharness_contract::v1::SessionEvent,
-        touch_session_activity: bool,
-    ) -> anyhow::Result<anyharness_contract::v1::SessionEventEnvelope> {
-        self.db.with_tx(|conn| {
-            let session_exists: bool = conn.query_row(
-                "SELECT EXISTS(SELECT 1 FROM sessions WHERE id = ?1)",
-                [session_id],
-                |row| row.get(0),
-            )?;
-            if !session_exists {
-                return Err(rusqlite::Error::QueryReturnedNoRows);
-            }
-
-            let seq: i64 = conn.query_row(
-                "SELECT COALESCE(MAX(seq), 0) + 1 FROM session_events WHERE session_id = ?1",
-                [session_id],
-                |row| row.get(0),
-            )?;
-            let timestamp = chrono::Utc::now().to_rfc3339();
-            let event_type = event.event_type().to_string();
-            let envelope = anyharness_contract::v1::SessionEventEnvelope {
-                session_id: session_id.to_string(),
-                seq,
-                timestamp: timestamp.clone(),
-                turn_id: None,
-                item_id: None,
-                event,
-            };
-            let persisted_event = sanitize_session_event_for_sqlite(&envelope.event);
-            let payload_json = serde_json::to_string(&persisted_event)
-                .map_err(|error| rusqlite::Error::ToSqlConversionFailure(Box::new(error)))?;
-            conn.execute(
-                "INSERT INTO session_events (
-                    session_id, seq, timestamp, event_type, turn_id, item_id, payload_json
-                 ) VALUES (?1, ?2, ?3, ?4, NULL, NULL, ?5)",
-                params![session_id, seq, timestamp, event_type, payload_json],
-            )?;
-            if touch_session_activity {
-                conn.execute(
-                    "UPDATE sessions SET updated_at = ?1 WHERE id = ?2",
-                    params![timestamp, session_id],
-                )?;
-            }
-            Ok(envelope)
-        })
-    }
-
-    pub fn append_raw_notification(
-        &self,
-        session_id: &str,
-        notification_kind: &str,
-        timestamp: &str,
-        payload_json: &str,
-    ) -> anyhow::Result<()> {
-        let sanitized_payload_json = sanitize_raw_notification_json_for_sqlite(payload_json);
-        self.db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO session_raw_notifications (session_id, seq, timestamp, notification_kind, payload_json)
-                 SELECT ?1, COALESCE(MAX(seq), 0) + 1, ?2, ?3, ?4
-                 FROM session_raw_notifications
-                 WHERE session_id = ?1",
-                params![session_id, timestamp, notification_kind, sanitized_payload_json],
-            )?;
-            Ok(())
-        })
-    }
-
-    pub fn list_events(&self, session_id: &str) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn
-                .prepare("SELECT * FROM session_events WHERE session_id = ?1 ORDER BY seq ASC")?;
-            let rows = stmt.query_map([session_id], |row| map_event(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_events_limited(
-        &self,
-        session_id: &str,
-        limit: i64,
-    ) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "WITH tail AS (
-                   SELECT seq, turn_id, item_id
-                   FROM session_events
-                   WHERE session_id = ?1
-                   ORDER BY seq DESC
-                   LIMIT ?2
-                 ),
-                 tail_turns AS (
-                   SELECT DISTINCT turn_id
-                   FROM tail
-                   WHERE turn_id IS NOT NULL
-                 ),
-                 tail_items AS (
-                   SELECT DISTINCT item_id
-                   FROM tail
-                   WHERE item_id IS NOT NULL
-                 )
-                 SELECT e.*
-                 FROM session_events e
-                 WHERE e.session_id = ?1
-                   AND (
-                     e.seq IN (SELECT seq FROM tail)
-                     OR (
-                       e.event_type = 'turn_started'
-                       AND e.turn_id IN (SELECT turn_id FROM tail_turns)
-                     )
-                     OR (
-                       e.event_type = 'item_started'
-                       AND e.item_id IN (SELECT item_id FROM tail_items)
-                     )
-                   )
-                 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map(params![session_id, limit], |row| map_event(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_events_for_latest_turns(
-        &self,
-        session_id: &str,
-        turn_limit: i64,
-        event_limit: i64,
-    ) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let turn_limit = turn_limit.max(1);
-            let event_limit = event_limit.max(1);
-            let mut turn_stmt = conn.prepare(
-                "SELECT turn_id, seq
-                 FROM session_events
-                 WHERE session_id = ?1
-                   AND event_type = 'turn_started'
-                   AND turn_id IS NOT NULL
-                 ORDER BY seq DESC
-                 LIMIT ?2",
-            )?;
-            let turn_rows = turn_stmt.query_map(params![session_id, turn_limit], |row| {
-                Ok((row.get::<_, String>("turn_id")?, row.get::<_, i64>("seq")?))
-            })?;
-            let turn_starts = turn_rows.collect::<rusqlite::Result<Vec<_>>>()?;
-
-            if turn_starts.is_empty() {
-                let mut stmt = conn.prepare(
-                    "SELECT *
-                     FROM (
-                       SELECT *
-                       FROM session_events
-                       WHERE session_id = ?1
-                       ORDER BY seq DESC
-                       LIMIT ?2
-                     )
-                     ORDER BY seq ASC",
-                )?;
-                let rows =
-                    stmt.query_map(params![session_id, event_limit], |row| map_event(row))?;
-                return rows.collect();
-            }
-
-            let mut selected_turn_count = turn_starts.len();
-            let mut cutoff_seq = turn_starts[selected_turn_count - 1].1;
-            loop {
-                let event_count: i64 = conn.query_row(
-                    "SELECT COUNT(*)
-                     FROM session_events
-                     WHERE session_id = ?1 AND seq >= ?2",
-                    params![session_id, cutoff_seq],
-                    |row| row.get(0),
-                )?;
-                if event_count <= event_limit || selected_turn_count <= 1 {
-                    break;
-                }
-                selected_turn_count -= 1;
-                cutoff_seq = turn_starts[selected_turn_count - 1].1;
-            }
-
-            let mut stmt = conn.prepare(
-                "SELECT *
-                 FROM session_events
-                 WHERE session_id = ?1 AND seq >= ?2
-                 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map(params![session_id, cutoff_seq], |row| map_event(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_events_before_limited(
-        &self,
-        session_id: &str,
-        before_seq: i64,
-        limit: i64,
-    ) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT *
-                 FROM (
-                   SELECT *
-                   FROM session_events
-                   WHERE session_id = ?1 AND seq < ?2
-                   ORDER BY seq DESC
-                   LIMIT ?3
-                 )
-                 ORDER BY seq ASC",
-            )?;
-            let rows =
-                stmt.query_map(params![session_id, before_seq, limit], |row| map_event(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_events_before_for_latest_turns(
-        &self,
-        session_id: &str,
-        before_seq: i64,
-        turn_limit: i64,
-        event_limit: i64,
-    ) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let turn_limit = turn_limit.max(1);
-            let event_limit = event_limit.max(1);
-            let mut turn_stmt = conn.prepare(
-                "SELECT turn_id, seq
-                 FROM session_events
-                 WHERE session_id = ?1
-                   AND seq < ?2
-                   AND event_type = 'turn_started'
-                   AND turn_id IS NOT NULL
-                 ORDER BY seq DESC
-                 LIMIT ?3",
-            )?;
-            let turn_rows = turn_stmt
-                .query_map(params![session_id, before_seq, turn_limit], |row| {
-                    Ok((row.get::<_, String>("turn_id")?, row.get::<_, i64>("seq")?))
-                })?;
-            let turn_starts = turn_rows.collect::<rusqlite::Result<Vec<_>>>()?;
-
-            if turn_starts.is_empty() {
-                let mut stmt = conn.prepare(
-                    "SELECT *
-                     FROM (
-                       SELECT *
-                       FROM session_events
-                       WHERE session_id = ?1 AND seq < ?2
-                       ORDER BY seq DESC
-                       LIMIT ?3
-                     )
-                     ORDER BY seq ASC",
-                )?;
-                let rows = stmt.query_map(params![session_id, before_seq, event_limit], |row| {
-                    map_event(row)
-                })?;
-                return rows.collect();
-            }
-
-            let mut selected_turn_count = turn_starts.len();
-            let mut cutoff_seq = turn_starts[selected_turn_count - 1].1;
-            loop {
-                let event_count: i64 = conn.query_row(
-                    "SELECT COUNT(*)
-                     FROM session_events
-                     WHERE session_id = ?1 AND seq >= ?2 AND seq < ?3",
-                    params![session_id, cutoff_seq, before_seq],
-                    |row| row.get(0),
-                )?;
-                if event_count <= event_limit || selected_turn_count <= 1 {
-                    break;
-                }
-                selected_turn_count -= 1;
-                cutoff_seq = turn_starts[selected_turn_count - 1].1;
-            }
-
-            let mut stmt = conn.prepare(
-                "SELECT *
-                 FROM session_events
-                 WHERE session_id = ?1 AND seq >= ?2 AND seq < ?3
-                 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map(params![session_id, cutoff_seq, before_seq], |row| {
-                map_event(row)
-            })?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_raw_notifications(
-        &self,
-        session_id: &str,
-    ) -> anyhow::Result<Vec<SessionRawNotificationRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_raw_notifications WHERE session_id = ?1 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map([session_id], map_raw_notification)?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_events_after(
-        &self,
-        session_id: &str,
-        after_seq: i64,
-    ) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_events
-                 WHERE session_id = ?1 AND seq > ?2
-                 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map(params![session_id, after_seq], |row| map_event(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_events_after_limited(
-        &self,
-        session_id: &str,
-        after_seq: i64,
-        limit: i64,
-    ) -> anyhow::Result<Vec<SessionEventRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT *
-                 FROM (
-                   SELECT *
-                   FROM session_events
-                   WHERE session_id = ?1 AND seq > ?2
-                   ORDER BY seq DESC
-                   LIMIT ?3
-                 )
-                 ORDER BY seq ASC",
-            )?;
-            let rows =
-                stmt.query_map(params![session_id, after_seq, limit], |row| map_event(row))?;
-            rows.collect()
-        })
-    }
-
-    pub fn list_raw_notifications_after(
-        &self,
-        session_id: &str,
-        after_seq: i64,
-    ) -> anyhow::Result<Vec<SessionRawNotificationRecord>> {
-        self.db.with_conn(|conn| {
-            let mut stmt = conn.prepare(
-                "SELECT * FROM session_raw_notifications
-                 WHERE session_id = ?1 AND seq > ?2
-                 ORDER BY seq ASC",
-            )?;
-            let rows = stmt.query_map(params![session_id, after_seq], map_raw_notification)?;
-            rows.collect()
-        })
-    }
-
-    pub fn last_event_seq(&self, session_id: &str) -> anyhow::Result<i64> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT COALESCE(MAX(seq), 0) FROM session_events WHERE session_id = ?1",
-                [session_id],
-                |row| row.get(0),
-            )
-        })
-    }
-
-    pub fn has_turn_started_event(&self, session_id: &str) -> anyhow::Result<bool> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT EXISTS(
-                     SELECT 1
-                     FROM session_events
-                     WHERE session_id = ?1 AND event_type = 'turn_started'
-                     LIMIT 1
-                 )",
-                [session_id],
-                |row| row.get(0),
-            )
-        })
-    }
-
-    pub fn count_turn_started_events(&self, session_id: &str) -> anyhow::Result<i64> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT COUNT(*)
-                 FROM session_events
-                 WHERE session_id = ?1 AND event_type = 'turn_started'",
-                [session_id],
-                |row| row.get(0),
-            )
-        })
-    }
-
-    pub fn has_terminal_turn_event(&self, session_id: &str) -> anyhow::Result<bool> {
-        self.db.with_conn(|conn| {
-            conn.query_row(
-                "SELECT EXISTS(
-                     SELECT 1
-                     FROM session_events
-                     WHERE session_id = ?1
-                       AND event_type IN ('turn_ended', 'error', 'session_ended')
-                     LIMIT 1
-                 )",
-                [session_id],
-                |row| row.get(0),
-            )
-        })
-    }
-
-    /// Find turns that have a `turn_started` but no corresponding `turn_ended`
-    /// (or `error` / `session_ended`) and close them with a synthetic
-    /// `turn_ended` event carrying `stop_reason: cancelled`. Returns the number
-    /// of turns repaired.
-    pub fn repair_unclosed_turns(&self, session_id: &str) -> anyhow::Result<u32> {
-        self.db.with_tx(|conn| {
-            // Find turn_ids that were started but never ended.
-            let mut stmt = conn.prepare(
-                "SELECT DISTINCT e.turn_id
-                 FROM session_events e
-                 WHERE e.session_id = ?1
-                   AND e.event_type = 'turn_started'
-                   AND e.turn_id IS NOT NULL
-                   AND NOT EXISTS (
-                     SELECT 1 FROM session_events e2
-                     WHERE e2.session_id = e.session_id
-                       AND e2.turn_id = e.turn_id
-                       AND e2.event_type IN ('turn_ended', 'error', 'session_ended')
-                   )",
-            )?;
-            let unclosed_turn_ids: Vec<String> = stmt
-                .query_map([session_id], |row| row.get(0))?
-                .collect::<Result<Vec<_>, _>>()?;
-
-            if unclosed_turn_ids.is_empty() {
-                return Ok(0);
-            }
-
-            let now = chrono::Utc::now().to_rfc3339();
-            let payload_json = r#"{"type":"turn_ended","stopReason":"cancelled"}"#;
-            let mut count = 0u32;
-
-            for turn_id in &unclosed_turn_ids {
-                let next_seq: i64 = conn.query_row(
-                    "SELECT COALESCE(MAX(seq), 0) + 1 FROM session_events WHERE session_id = ?1",
-                    [session_id],
-                    |row| row.get(0),
-                )?;
-
-                conn.execute(
-                    "INSERT INTO session_events (session_id, seq, timestamp, event_type, turn_id, item_id, payload_json)
-                     VALUES (?1, ?2, ?3, ?4, ?5, NULL, ?6)",
-                    params![session_id, next_seq, now, "turn_ended", turn_id, payload_json],
-                )?;
-
-                tracing::info!(
-                    session_id = %session_id,
-                    turn_id = %turn_id,
-                    seq = next_seq,
-                    "repaired unclosed turn with synthetic turn_ended"
-                );
-                count += 1;
-            }
-
-            Ok(count)
-        })
-    }
-}
-
-pub(crate) fn delete_session_in_tx(conn: &rusqlite::Connection, id: &str) -> rusqlite::Result<()> {
-    // Several durable tables still reference sessions without database-level
-    // cascade rules, so session deletion must clear dependent rows explicitly.
-    conn.execute("DELETE FROM cowork_threads WHERE session_id = ?1", [id])?;
-    conn.execute(
-        "DELETE FROM cowork_managed_workspaces WHERE parent_session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_background_work WHERE session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM review_feedback_jobs
-         WHERE review_run_id IN (
-            SELECT id FROM review_runs
-            WHERE parent_session_id = ?1
-         )",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM review_run_candidate_plans
-         WHERE review_run_id IN (
-            SELECT id FROM review_runs
-            WHERE parent_session_id = ?1
-         )",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM review_assignments
-         WHERE review_run_id IN (
-            SELECT id FROM review_runs
-            WHERE parent_session_id = ?1
-         ) OR reviewer_session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM review_rounds
-         WHERE review_run_id IN (
-            SELECT id FROM review_runs
-            WHERE parent_session_id = ?1
-         )",
-        [id],
-    )?;
-    conn.execute("DELETE FROM review_runs WHERE parent_session_id = ?1", [id])?;
-    conn.execute(
-        "DELETE FROM session_link_wake_schedules
-         WHERE session_link_id IN (
-            SELECT id FROM session_links
-            WHERE parent_session_id = ?1 OR child_session_id = ?1
-         )",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_link_completions
-         WHERE session_link_id IN (
-            SELECT id FROM session_links
-            WHERE parent_session_id = ?1 OR child_session_id = ?1
-         )",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_links WHERE parent_session_id = ?1 OR child_session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_pending_prompts WHERE session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_prompt_attachments WHERE session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_pending_config_changes WHERE session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_live_config_snapshots WHERE session_id = ?1",
-        [id],
-    )?;
-    conn.execute(
-        "DELETE FROM session_raw_notifications WHERE session_id = ?1",
-        [id],
-    )?;
-    conn.execute("DELETE FROM session_events WHERE session_id = ?1", [id])?;
-    conn.execute("DELETE FROM sessions WHERE id = ?1", [id])?;
-    Ok(())
-}
-
-fn map_session(row: &rusqlite::Row) -> rusqlite::Result<SessionRecord> {
-    let id: String = row.get("id")?;
-    let origin_json: Option<String> = row.get("origin_json")?;
-    Ok(SessionRecord {
-        id: id.clone(),
-        workspace_id: row.get("workspace_id")?,
-        agent_kind: row.get("agent_kind")?,
-        native_session_id: row.get("native_session_id")?,
-        requested_model_id: row.get("requested_model_id")?,
-        current_model_id: row.get("current_model_id")?,
-        requested_mode_id: row.get("requested_mode_id")?,
-        current_mode_id: row.get("current_mode_id")?,
-        title: row.get("title")?,
-        thinking_level_id: row.get("thinking_level_id")?,
-        thinking_budget_tokens: row.get("thinking_budget_tokens")?,
-        status: row.get("status")?,
-        created_at: row.get("created_at")?,
-        updated_at: row.get("updated_at")?,
-        last_prompt_at: row.get("last_prompt_at")?,
-        closed_at: row.get("closed_at")?,
-        dismissed_at: row.get("dismissed_at")?,
-        mcp_bindings_ciphertext: row.get("mcp_bindings_ciphertext")?,
-        mcp_binding_summaries_json: row.get("mcp_binding_summaries_json")?,
-        mcp_binding_policy: SessionMcpBindingPolicy::parse(
-            &row.get::<_, String>("mcp_binding_policy")?,
-        ),
-        system_prompt_append: row.get("system_prompt_append")?,
-        subagents_enabled: row.get::<_, i64>("subagents_enabled")? != 0,
-        action_capabilities_json: row.get("action_capabilities_json")?,
-        origin: decode_origin_json("sessions", &id, origin_json),
-    })
-}
-
-fn map_live_config_snapshot(
-    row: &rusqlite::Row,
-) -> rusqlite::Result<SessionLiveConfigSnapshotRecord> {
-    Ok(SessionLiveConfigSnapshotRecord {
-        session_id: row.get("session_id")?,
-        source_seq: row.get("source_seq")?,
-        raw_config_options_json: row.get("raw_config_options_json")?,
-        normalized_controls_json: row.get("normalized_controls_json")?,
-        prompt_capabilities_json: row.get("prompt_capabilities_json")?,
-        updated_at: row.get("updated_at")?,
-    })
-}
-
-fn map_event(row: &rusqlite::Row) -> rusqlite::Result<SessionEventRecord> {
-    Ok(SessionEventRecord {
-        id: row.get("id")?,
-        session_id: row.get("session_id")?,
-        seq: row.get("seq")?,
-        timestamp: row.get("timestamp")?,
-        event_type: row.get("event_type")?,
-        turn_id: row.get("turn_id")?,
-        item_id: row.get("item_id")?,
-        payload_json: row.get("payload_json")?,
-    })
-}
-
-fn map_raw_notification(row: &rusqlite::Row) -> rusqlite::Result<SessionRawNotificationRecord> {
-    Ok(SessionRawNotificationRecord {
-        id: row.get("id")?,
-        session_id: row.get("session_id")?,
-        seq: row.get("seq")?,
-        timestamp: row.get("timestamp")?,
-        notification_kind: row.get("notification_kind")?,
-        payload_json: row.get("payload_json")?,
-    })
-}
-
-fn map_pending_config_change(row: &rusqlite::Row) -> rusqlite::Result<PendingConfigChangeRecord> {
-    Ok(PendingConfigChangeRecord {
-        session_id: row.get("session_id")?,
-        config_id: row.get("config_id")?,
-        value: row.get("value")?,
-        queued_at: row.get("queued_at")?,
-    })
-}
-
-fn map_pending_prompt(row: &rusqlite::Row) -> rusqlite::Result<PendingPromptRecord> {
-    Ok(PendingPromptRecord {
-        session_id: row.get("session_id")?,
-        seq: row.get("seq")?,
-        prompt_id: row.get("prompt_id")?,
-        text: row.get("text")?,
-        blocks_json: row.get("blocks_json")?,
-        provenance_json: row.get("provenance_json")?,
-        queued_at: row.get("queued_at")?,
-    })
-}
-
-fn map_prompt_attachment(row: &rusqlite::Row) -> rusqlite::Result<PromptAttachmentRecord> {
-    let state: String = row.get("state")?;
-    let kind: String = row.get("kind")?;
-    let source: String = row.get("source")?;
-    Ok(PromptAttachmentRecord {
-        attachment_id: row.get("attachment_id")?,
-        session_id: row.get("session_id")?,
-        state: PromptAttachmentState::parse(&state),
-        kind: PromptAttachmentKind::parse(&kind),
-        source: PromptAttachmentSource::parse(&source),
-        mime_type: row.get("mime_type")?,
-        display_name: row.get("display_name")?,
-        source_uri: row.get("source_uri")?,
-        storage_path: row.get("storage_path")?,
-        size_bytes: row.get("size_bytes")?,
-        sha256: row.get("sha256")?,
-        created_at: row.get("created_at")?,
-        updated_at: row.get("updated_at")?,
-    })
-}
-
-fn insert_session_row(conn: &rusqlite::Connection, record: &SessionRecord) -> rusqlite::Result<()> {
-    let origin_json = encode_origin_json(&record.origin)?;
-    conn.execute(
-        "INSERT INTO sessions (id, workspace_id, agent_kind, native_session_id,
-         requested_model_id, current_model_id, requested_mode_id, current_mode_id,
-         title, thinking_level_id, thinking_budget_tokens, status, created_at,
-         updated_at, last_prompt_at, closed_at, dismissed_at, mcp_bindings_ciphertext,
-         mcp_binding_summaries_json, mcp_binding_policy, system_prompt_append,
-         subagents_enabled, action_capabilities_json, origin_json)
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, ?16, ?17, ?18, ?19, ?20, ?21, ?22, ?23, ?24)",
-        params![
-            record.id,
-            record.workspace_id,
-            record.agent_kind,
-            record.native_session_id,
-            record.requested_model_id,
-            record.current_model_id,
-            record.requested_mode_id,
-            record.current_mode_id,
-            record.title,
-            record.thinking_level_id,
-            record.thinking_budget_tokens,
-            record.status,
-            record.created_at,
-            record.updated_at,
-            record.last_prompt_at,
-            record.closed_at,
-            record.dismissed_at,
-            record.mcp_bindings_ciphertext,
-            record.mcp_binding_summaries_json,
-            record.mcp_binding_policy.as_str(),
-            record.system_prompt_append,
-            if record.subagents_enabled { 1 } else { 0 },
-            record.action_capabilities_json,
-            origin_json,
-        ],
-    )?;
-    Ok(())
-}
-
-fn insert_session_link_row(
-    conn: &rusqlite::Connection,
-    record: &SessionLinkRecord,
-) -> rusqlite::Result<()> {
-    conn.execute(
-        "INSERT INTO session_links (
-            id, relation, parent_session_id, child_session_id, workspace_relation,
-            label, created_by_turn_id, created_by_tool_call_id, created_at
-         ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
-        params![
-            record.id,
-            record.relation.as_str(),
-            record.parent_session_id,
-            record.child_session_id,
-            record.workspace_relation.as_str(),
-            record.label,
-            record.created_by_turn_id,
-            record.created_by_tool_call_id,
-            record.created_at,
-        ],
-    )?;
-    Ok(())
-}
-
-fn upsert_live_config_snapshot_row(
-    conn: &rusqlite::Connection,
-    record: &SessionLiveConfigSnapshotRecord,
-) -> rusqlite::Result<()> {
-    conn.execute(
-        "INSERT INTO session_live_config_snapshots (
-            session_id, source_seq, raw_config_options_json, normalized_controls_json,
-            prompt_capabilities_json, updated_at
-         ) VALUES (?1, ?2, ?3, ?4, ?5, ?6)
-         ON CONFLICT(session_id) DO UPDATE SET
-            source_seq = excluded.source_seq,
-            raw_config_options_json = excluded.raw_config_options_json,
-            normalized_controls_json = excluded.normalized_controls_json,
-            prompt_capabilities_json = excluded.prompt_capabilities_json,
-            updated_at = excluded.updated_at",
-        params![
-            record.session_id,
-            record.source_seq,
-            record.raw_config_options_json,
-            record.normalized_controls_json,
-            record.prompt_capabilities_json,
-            record.updated_at,
-        ],
-    )?;
-    Ok(())
-}
-
-fn upsert_pending_config_change_row(
-    conn: &rusqlite::Connection,
-    record: &PendingConfigChangeRecord,
-) -> rusqlite::Result<()> {
-    conn.execute(
-        "INSERT INTO session_pending_config_changes (session_id, config_id, value, queued_at)
-         VALUES (?1, ?2, ?3, ?4)
-         ON CONFLICT(session_id, config_id) DO UPDATE SET
-            value = excluded.value,
-            queued_at = excluded.queued_at",
-        params![
-            record.session_id,
-            record.config_id,
-            record.value,
-            record.queued_at
-        ],
-    )?;
-    Ok(())
-}
-
-fn insert_pending_prompt_row(
-    conn: &rusqlite::Connection,
-    record: &PendingPromptRecord,
-) -> rusqlite::Result<()> {
-    conn.execute(
-        "INSERT INTO session_pending_prompts (
-            session_id, seq, prompt_id, text, blocks_json, queued_at, provenance_json
-         ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-        params![
-            record.session_id,
-            record.seq,
-            record.prompt_id,
-            record.text,
-            record.blocks_json,
-            record.queued_at,
-            record.provenance_json,
-        ],
-    )?;
-    Ok(())
-}
-
-fn insert_prompt_attachment_row(
-    conn: &rusqlite::Connection,
-    record: &PromptAttachmentRecord,
-) -> rusqlite::Result<()> {
-    conn.execute(
-        "INSERT INTO session_prompt_attachments (
-            attachment_id, session_id, state, kind, source, mime_type, display_name, source_uri,
-            storage_path, size_bytes, sha256, content, created_at, updated_at
-         ) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14)
-         ON CONFLICT(attachment_id) DO UPDATE SET
-            session_id = excluded.session_id,
-            state = excluded.state,
-            kind = excluded.kind,
-            source = excluded.source,
-            mime_type = excluded.mime_type,
-            display_name = excluded.display_name,
-            source_uri = excluded.source_uri,
-            storage_path = excluded.storage_path,
-            size_bytes = excluded.size_bytes,
-            sha256 = excluded.sha256,
-            content = excluded.content,
-            updated_at = excluded.updated_at",
-        params![
-            record.attachment_id,
-            record.session_id,
-            record.state.as_str(),
-            record.kind.as_str(),
-            record.source.as_str(),
-            record.mime_type,
-            record.display_name,
-            record.source_uri,
-            record.storage_path,
-            record.size_bytes,
-            record.sha256,
-            Vec::<u8>::new(),
-            record.created_at,
-            record.updated_at,
-        ],
-    )?;
-    Ok(())
-}
-
-fn insert_event_row(
-    conn: &rusqlite::Connection,
-    record: &SessionEventRecord,
-) -> rusqlite::Result<()> {
-    let payload_json = sanitize_session_event_payload_json(&record.payload_json)?;
-    conn.execute(
-        "INSERT INTO session_events (session_id, seq, timestamp, event_type, turn_id, item_id, payload_json)
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-        params![
-            record.session_id,
-            record.seq,
-            record.timestamp,
-            record.event_type,
-            record.turn_id,
-            record.item_id,
-            payload_json,
-        ],
-    )?;
-    Ok(())
-}
-
-fn sanitize_session_event_payload_json(payload_json: &str) -> rusqlite::Result<String> {
-    let Ok(event) = serde_json::from_str::<anyharness_contract::v1::SessionEvent>(payload_json)
-    else {
-        return Ok(payload_json.to_string());
-    };
-    let sanitized = sanitize_session_event_for_sqlite(&event);
-    serde_json::to_string(&sanitized)
-        .map_err(|error| rusqlite::Error::ToSqlConversionFailure(Box::new(error)))
-}
-
-fn insert_raw_notification_row(
-    conn: &rusqlite::Connection,
-    record: &SessionRawNotificationRecord,
-) -> rusqlite::Result<()> {
-    let payload_json = sanitize_raw_notification_json_for_sqlite(&record.payload_json);
-    conn.execute(
-        "INSERT INTO session_raw_notifications (session_id, seq, timestamp, notification_kind, payload_json)
-         VALUES (?1, ?2, ?3, ?4, ?5)",
-        params![
-            record.session_id,
-            record.seq,
-            record.timestamp,
-            record.notification_kind,
-            payload_json,
-        ],
-    )?;
-    Ok(())
-}
-
-fn map_background_work(row: &rusqlite::Row) -> rusqlite::Result<SessionBackgroundWorkRecord> {
-    let tracker_kind: String = row.get("tracker_kind")?;
-    let state: String = row.get("state")?;
-
-    Ok(SessionBackgroundWorkRecord {
-        session_id: row.get("session_id")?,
-        tool_call_id: row.get("tool_call_id")?,
-        turn_id: row.get("turn_id")?,
-        tracker_kind: SessionBackgroundWorkTrackerKind::parse(&tracker_kind),
-        source_agent_kind: row.get("source_agent_kind")?,
-        agent_id: row.get("agent_id")?,
-        output_file: row.get("output_file")?,
-        state: SessionBackgroundWorkState::parse(&state),
-        created_at: row.get("created_at")?,
-        updated_at: row.get("updated_at")?,
-        launched_at: row.get("launched_at")?,
-        last_activity_at: row.get("last_activity_at")?,
-        completed_at: row.get("completed_at")?,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::origin::OriginContext;
-    use crate::persistence::Db;
-    use crate::sessions::links::model::SessionLinkWorkspaceRelation;
-    use crate::sessions::prompt::{PromptPayload, PromptProvenance};
-
-    fn count_rows(db: &Db, table: &str, session_id: &str) -> i64 {
-        let sql = format!("SELECT COUNT(*) FROM {table} WHERE session_id = ?1");
-        db.with_conn(|conn| conn.query_row(&sql, [session_id], |row| row.get(0)))
-            .expect("count rows")
-    }
-
-    fn count_all_rows(db: &Db, table: &str) -> i64 {
-        let sql = format!("SELECT COUNT(*) FROM {table}");
-        db.with_conn(|conn| conn.query_row(&sql, [], |row| row.get(0)))
-            .expect("count all rows")
-    }
-
-    fn seed_workspace(db: &Db) {
-        db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO workspaces (id, kind, path, source_repo_root_path, created_at, updated_at)
-                 VALUES (?1, 'repo', '/tmp/workspace', '/tmp/workspace', ?2, ?2)",
-                params!["workspace-1", "2026-03-25T00:00:00Z"],
-            )?;
-            Ok(())
-        })
-        .expect("seed workspace");
-    }
-
-    fn session_record() -> SessionRecord {
-        SessionRecord {
-            id: "session-1".to_string(),
-            workspace_id: "workspace-1".to_string(),
-            agent_kind: "claude".to_string(),
-            native_session_id: Some("native-1".to_string()),
-            requested_model_id: Some("default".to_string()),
-            current_model_id: Some("default".to_string()),
-            requested_mode_id: Some("default".to_string()),
-            current_mode_id: Some("default".to_string()),
-            title: Some("Fix auth refresh".to_string()),
-            thinking_level_id: None,
-            thinking_budget_tokens: Some(16_000),
-            status: "idle".to_string(),
-            created_at: "2026-03-25T00:00:00Z".to_string(),
-            updated_at: "2026-03-25T00:00:00Z".to_string(),
-            last_prompt_at: None,
-            closed_at: None,
-            dismissed_at: None,
-            mcp_bindings_ciphertext: None,
-            mcp_binding_summaries_json: None,
-            mcp_binding_policy: crate::sessions::model::SessionMcpBindingPolicy::InheritWorkspace,
-            system_prompt_append: None,
-            subagents_enabled: true,
-            action_capabilities_json: None,
-            origin: None,
-        }
-    }
-
-    fn fork_link_record(
-        id: &str,
-        parent_session_id: &str,
-        child_session_id: &str,
-    ) -> SessionLinkRecord {
-        SessionLinkRecord {
-            id: id.to_string(),
-            relation: SessionLinkRelation::Fork,
-            parent_session_id: parent_session_id.to_string(),
-            child_session_id: child_session_id.to_string(),
-            workspace_relation: SessionLinkWorkspaceRelation::SameWorkspace,
-            label: None,
-            created_by_turn_id: None,
-            created_by_tool_call_id: None,
-            created_at: "2026-03-25T00:00:00Z".to_string(),
-        }
-    }
-
-    #[test]
-    fn insert_session_with_link_rolls_back_child_when_link_insert_fails() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        let mut parent = session_record();
-        parent.id = "parent-session".to_string();
-        store.insert(&parent).expect("insert parent session");
-
-        let mut child = session_record();
-        child.id = "child-session".to_string();
-        child.native_session_id = Some("native-child".to_string());
-        let first_link = fork_link_record("duplicate-link", "parent-session", "child-session");
-        store
-            .insert_session_with_link(&child, &first_link)
-            .expect("insert first child and link");
-
-        let mut second_child = session_record();
-        second_child.id = "second-child".to_string();
-        second_child.native_session_id = Some("native-second-child".to_string());
-        let duplicate_link = fork_link_record("duplicate-link", "parent-session", "second-child");
-
-        store
-            .insert_session_with_link(&second_child, &duplicate_link)
-            .expect_err("duplicate link id rejects transaction");
-
-        assert!(store
-            .find_by_id("second-child")
-            .expect("find second child")
-            .is_none());
-    }
-
-    #[test]
-    fn insert_fork_session_with_link_snapshots_parent_events() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db.clone());
-        let mut parent = session_record();
-        parent.id = "parent-session".to_string();
-        store.insert(&parent).expect("insert parent session");
-        store
-            .append_event(&SessionEventRecord {
-                id: 0,
-                session_id: "parent-session".to_string(),
-                seq: 1,
-                timestamp: "2026-03-25T00:01:00Z".to_string(),
-                event_type: "turn_started".to_string(),
-                turn_id: Some("turn-1".to_string()),
-                item_id: None,
-                payload_json: r#"{"type":"turn_started"}"#.to_string(),
-            })
-            .expect("append parent turn");
-        store
-            .append_event(&SessionEventRecord {
-                id: 0,
-                session_id: "parent-session".to_string(),
-                seq: 2,
-                timestamp: "2026-03-25T00:01:01Z".to_string(),
-                event_type: "item_started".to_string(),
-                turn_id: Some("turn-1".to_string()),
-                item_id: Some("item-1".to_string()),
-                payload_json: r#"{"type":"item_started","item":{"kind":"user_message","status":"completed","contentParts":[{"type":"text","text":"hi"}]}}"#.to_string(),
-            })
-            .expect("append parent item");
-        store
-            .append_raw_notification(
-                "parent-session",
-                "agent_message_chunk",
-                "2026-03-25T00:01:02Z",
-                r#"{"chunk":"raw"}"#,
-            )
-            .expect("append parent raw notification");
-
-        let mut child = session_record();
-        child.id = "child-session".to_string();
-        child.native_session_id = None;
-        let link = fork_link_record("fork-link", "parent-session", "child-session");
-        let copied = store
-            .insert_fork_session_with_link_and_event_snapshot(&child, &link)
-            .expect("insert fork child with event snapshot");
-
-        assert_eq!(copied, 2);
-        assert_eq!(store.next_event_seq("child-session").expect("next seq"), 3);
-        let events = store
-            .list_events("child-session")
-            .expect("list child events");
-        assert_eq!(events.len(), 2);
-        assert_eq!(events[0].seq, 1);
-        assert_eq!(events[0].event_type, "turn_started");
-        assert_eq!(events[0].turn_id.as_deref(), Some("turn-1"));
-        assert_eq!(events[1].seq, 2);
-        assert_eq!(events[1].item_id.as_deref(), Some("item-1"));
-        assert_eq!(
-            events[1].payload_json,
-            r#"{"type":"item_started","item":{"kind":"user_message","status":"completed","contentParts":[{"type":"text","text":"hi"}]}}"#
-        );
-        assert!(
-            store
-                .list_raw_notifications("child-session")
-                .expect("list child raw notifications")
-                .is_empty(),
-            "fork transcript snapshots must not copy raw ACP notifications"
-        );
-    }
-
-    #[test]
-    fn stores_and_loads_session_origin() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        let mut record = session_record();
-        record.origin = Some(OriginContext::cowork());
-
-        store.insert(&record).expect("insert session");
-        let stored = store
-            .find_by_id("session-1")
-            .expect("find session")
-            .expect("session record");
-
-        assert_eq!(stored.origin, Some(OriginContext::cowork()));
-    }
-
-    #[test]
-    fn stores_and_loads_thinking_budget_tokens() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        let record = session_record();
-
-        store.insert(&record).expect("insert session");
-        let stored = store
-            .find_by_id("session-1")
-            .expect("find session")
-            .expect("session record");
-
-        assert_eq!(stored.thinking_budget_tokens, Some(16_000));
-        assert_eq!(stored.title.as_deref(), Some("Fix auth refresh"));
-    }
-
-    #[test]
-    fn detects_when_a_session_has_started_a_turn() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        let record = session_record();
-        store.insert(&record).expect("insert session");
-
-        assert!(!store
-            .has_turn_started_event("session-1")
-            .expect("check empty turn history"));
-
-        store
-            .append_event(&SessionEventRecord {
-                id: 0,
-                session_id: "session-1".to_string(),
-                seq: 1,
-                timestamp: "2026-03-25T00:01:00Z".to_string(),
-                event_type: "turn_started".to_string(),
-                turn_id: Some("turn-1".to_string()),
-                item_id: None,
-                payload_json: r#"{"type":"turn_started"}"#.to_string(),
-            })
-            .expect("append turn_started");
-
-        assert!(store
-            .has_turn_started_event("session-1")
-            .expect("check populated turn history"));
-    }
-
-    #[test]
-    fn limited_event_reads_return_newest_events_in_ascending_order() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-
-        for seq in 1..=5 {
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq,
-                    timestamp: format!("2026-03-25T00:01:0{seq}Z"),
-                    event_type: "turn_started".to_string(),
-                    turn_id: Some(format!("turn-{seq}")),
-                    item_id: None,
-                    payload_json: r#"{"type":"turn_started"}"#.to_string(),
-                })
-                .expect("append event");
-        }
-
-        let tail = store
-            .list_events_limited("session-1", 2)
-            .expect("list limited events");
-        assert_eq!(
-            tail.iter().map(|event| event.seq).collect::<Vec<_>>(),
-            vec![4, 5]
-        );
-
-        let filtered_tail = store
-            .list_events_after_limited("session-1", 2, 2)
-            .expect("list limited events after seq");
-        assert_eq!(
-            filtered_tail
-                .iter()
-                .map(|event| event.seq)
-                .collect::<Vec<_>>(),
-            vec![4, 5],
-        );
-    }
-
-    #[test]
-    fn limited_event_reads_include_tail_turn_and_item_start_context() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-
-        store
-            .append_event(&SessionEventRecord {
-                id: 0,
-                session_id: "session-1".to_string(),
-                seq: 1,
-                timestamp: "2026-03-25T00:01:01Z".to_string(),
-                event_type: "turn_started".to_string(),
-                turn_id: Some("turn-1".to_string()),
-                item_id: None,
-                payload_json: r#"{"type":"turn_started"}"#.to_string(),
-            })
-            .expect("append turn start");
-
-        for seq in 2..=4 {
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq,
-                    timestamp: format!("2026-03-25T00:01:0{seq}Z"),
-                    event_type: if seq == 2 { "item_started" } else { "item_delta" }.to_string(),
-                    turn_id: Some("turn-1".to_string()),
-                    item_id: Some("item-1".to_string()),
-                    payload_json: if seq == 2 {
-                        r#"{"type":"item_started","item":{"kind":"assistant_message","status":"in_progress","sourceAgentKind":"codex","contentParts":[]}}"#
-                    } else {
-                        r#"{"type":"item_delta","delta":{"appendText":"old"}}"#
-                    }
-                    .to_string(),
-                })
-                .expect("append older item event");
-        }
-
-        for seq in 5..=7 {
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq,
-                    timestamp: format!("2026-03-25T00:01:0{seq}Z"),
-                    event_type: if seq == 5 { "item_started" } else { "item_delta" }.to_string(),
-                    turn_id: Some("turn-1".to_string()),
-                    item_id: Some("item-2".to_string()),
-                    payload_json: if seq == 5 {
-                        r#"{"type":"item_started","item":{"kind":"assistant_message","status":"in_progress","sourceAgentKind":"codex","contentParts":[]}}"#
-                    } else {
-                        r#"{"type":"item_delta","delta":{"appendText":"new"}}"#
-                    }
-                    .to_string(),
-                })
-                .expect("append recent item event");
-        }
-
-        let tail = store
-            .list_events_limited("session-1", 2)
-            .expect("list limited events");
-
-        assert_eq!(
-            tail.iter().map(|event| event.seq).collect::<Vec<_>>(),
-            vec![1, 5, 6, 7],
-        );
-    }
-
-    #[test]
-    fn latest_turn_reads_return_complete_recent_turns() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-
-        for turn in 1..=3 {
-            let start_seq = (turn - 1) * 2 + 1;
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq: start_seq,
-                    timestamp: format!("2026-03-25T00:01:{start_seq:02}Z"),
-                    event_type: "turn_started".to_string(),
-                    turn_id: Some(format!("turn-{turn}")),
-                    item_id: None,
-                    payload_json: r#"{"type":"turn_started"}"#.to_string(),
-                })
-                .expect("append turn_started");
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq: start_seq + 1,
-                    timestamp: format!("2026-03-25T00:01:{:02}Z", start_seq + 1),
-                    event_type: "turn_ended".to_string(),
-                    turn_id: Some(format!("turn-{turn}")),
-                    item_id: None,
-                    payload_json: r#"{"type":"turn_ended","stopReason":"end_turn"}"#.to_string(),
-                })
-                .expect("append turn_ended");
-        }
-
-        let tail = store
-            .list_events_for_latest_turns("session-1", 2, 100)
-            .expect("list latest turns");
-
-        assert_eq!(
-            tail.iter().map(|event| event.seq).collect::<Vec<_>>(),
-            vec![3, 4, 5, 6],
-        );
-    }
-
-    #[test]
-    fn latest_turn_reads_reduce_window_to_event_budget() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-
-        for turn in 1..=3 {
-            let start_seq = (turn - 1) * 3 + 1;
-            for offset in 0..3 {
-                let seq = start_seq + offset;
-                store
-                    .append_event(&SessionEventRecord {
-                        id: 0,
-                        session_id: "session-1".to_string(),
-                        seq,
-                        timestamp: format!("2026-03-25T00:01:{seq:02}Z"),
-                        event_type: if offset == 0 {
-                            "turn_started"
-                        } else {
-                            "item_completed"
-                        }
-                        .to_string(),
-                        turn_id: Some(format!("turn-{turn}")),
-                        item_id: Some(format!("item-{turn}-{offset}")),
-                        payload_json: if offset == 0 {
-                            r#"{"type":"turn_started"}"#
-                        } else {
-                            r#"{"type":"item_completed","item":{"kind":"assistant_message","status":"completed","sourceAgentKind":"codex","contentParts":[]}}"#
-                        }
-                        .to_string(),
-                    })
-                    .expect("append event");
-            }
-        }
-
-        let tail = store
-            .list_events_for_latest_turns("session-1", 3, 4)
-            .expect("list budgeted latest turns");
-
-        assert_eq!(
-            tail.iter().map(|event| event.seq).collect::<Vec<_>>(),
-            vec![7, 8, 9],
-        );
-    }
-
-    #[test]
-    fn older_turn_reads_return_complete_page_before_cutoff() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-
-        for turn in 1..=5 {
-            let start_seq = (turn - 1) * 2 + 1;
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq: start_seq,
-                    timestamp: format!("2026-03-25T00:01:{start_seq:02}Z"),
-                    event_type: "turn_started".to_string(),
-                    turn_id: Some(format!("turn-{turn}")),
-                    item_id: None,
-                    payload_json: r#"{"type":"turn_started"}"#.to_string(),
-                })
-                .expect("append turn_started");
-            store
-                .append_event(&SessionEventRecord {
-                    id: 0,
-                    session_id: "session-1".to_string(),
-                    seq: start_seq + 1,
-                    timestamp: format!("2026-03-25T00:01:{:02}Z", start_seq + 1),
-                    event_type: "turn_ended".to_string(),
-                    turn_id: Some(format!("turn-{turn}")),
-                    item_id: None,
-                    payload_json: r#"{"type":"turn_ended","stopReason":"end_turn"}"#.to_string(),
-                })
-                .expect("append turn_ended");
-        }
-
-        let older = store
-            .list_events_before_for_latest_turns("session-1", 7, 2, 100)
-            .expect("list older turns");
-
-        assert_eq!(
-            older.iter().map(|event| event.seq).collect::<Vec<_>>(),
-            vec![3, 4, 5, 6],
-        );
-    }
-
-    #[test]
-    fn pending_prompt_preserves_internal_provenance_through_load_edit_and_drain() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-        let payload = PromptPayload {
-            text_summary: "delegate this".to_string(),
-            blocks: vec![],
-            provenance: Some(PromptProvenance::AgentSession {
-                source_session_id: "parent-session".to_string(),
-                session_link_id: Some("link-1".to_string()),
-                label: Some("Parent agent".to_string()),
-            }),
-        };
-
-        let inserted = store
-            .insert_pending_prompt_payload("session-1", &payload, Some("prompt-1"))
-            .expect("insert pending prompt");
-        assert!(inserted.provenance_json.is_some());
-
-        let loaded = store
-            .find_pending_prompt("session-1", inserted.seq)
-            .expect("find prompt")
-            .expect("prompt exists");
-        assert_eq!(loaded.prompt_payload().provenance, payload.provenance);
-
-        store
-            .update_pending_prompt_text("session-1", inserted.seq, "edited")
-            .expect("edit prompt");
-        let edited = store
-            .find_pending_prompt("session-1", inserted.seq)
-            .expect("find edited prompt")
-            .expect("edited prompt exists");
-        assert_eq!(edited.text, "edited");
-        assert_eq!(edited.prompt_payload().provenance, payload.provenance);
-
-        let drained = store
-            .delete_pending_prompt_record("session-1", inserted.seq)
-            .expect("drain prompt")
-            .expect("drained prompt exists");
-        assert_eq!(drained.prompt_payload().provenance, payload.provenance);
-    }
-
-    #[test]
-    fn delete_session_removes_dependent_rows() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db.clone());
-        let record = session_record();
-        store.insert(&record).expect("insert session");
-        let mut child = session_record();
-        child.id = "session-child".to_string();
-        store.insert(&child).expect("insert child session");
-
-        store
-            .append_event(&SessionEventRecord {
-                id: 0,
-                session_id: "session-1".to_string(),
-                seq: 1,
-                timestamp: "2026-03-25T00:01:00Z".to_string(),
-                event_type: "turn_started".to_string(),
-                turn_id: Some("turn-1".to_string()),
-                item_id: None,
-                payload_json: r#"{"type":"turn_started"}"#.to_string(),
-            })
-            .expect("append event");
-        store
-            .append_raw_notification(
-                "session-1",
-                "agent_message_chunk",
-                "2026-03-25T00:01:01Z",
-                r#"{"kind":"agent_message_chunk"}"#,
-            )
-            .expect("append raw notification");
-        store
-            .upsert_live_config_snapshot(&SessionLiveConfigSnapshotRecord {
-                session_id: "session-1".to_string(),
-                source_seq: 1,
-                raw_config_options_json: "{}".to_string(),
-                normalized_controls_json: "{}".to_string(),
-                prompt_capabilities_json: None,
-                updated_at: "2026-03-25T00:01:02Z".to_string(),
-            })
-            .expect("upsert snapshot");
-        store
-            .upsert_pending_config_change(&PendingConfigChangeRecord {
-                session_id: "session-1".to_string(),
-                config_id: "model".to_string(),
-                value: "\"opus\"".to_string(),
-                queued_at: "2026-03-25T00:01:03Z".to_string(),
-            })
-            .expect("insert pending config change");
-        store
-            .insert_pending_prompt("session-1", "finish cleanup", Some("prompt-1"))
-            .expect("insert pending prompt");
-        store
-            .upsert_or_refresh_pending_background_work(&SessionBackgroundWorkRecord {
-                session_id: "session-1".to_string(),
-                tool_call_id: "tool-1".to_string(),
-                turn_id: "turn-1".to_string(),
-                tracker_kind: SessionBackgroundWorkTrackerKind::ClaudeAsyncAgent,
-                source_agent_kind: "claude".to_string(),
-                agent_id: Some("agent-1".to_string()),
-                output_file: "/tmp/agent.output".to_string(),
-                state: SessionBackgroundWorkState::Pending,
-                created_at: "2026-03-25T00:01:04Z".to_string(),
-                updated_at: "2026-03-25T00:01:04Z".to_string(),
-                launched_at: "2026-03-25T00:01:04Z".to_string(),
-                last_activity_at: "2026-03-25T00:01:04Z".to_string(),
-                completed_at: None,
-            })
-            .expect("insert background work");
-        db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO session_links (
-                    id, relation, parent_session_id, child_session_id, workspace_relation,
-                    created_at
-                 ) VALUES ('link-1', 'subagent', 'session-1', 'session-child', 'same_workspace', ?1)",
-                ["2026-03-25T00:01:05Z"],
-            )?;
-            conn.execute(
-                "INSERT INTO session_link_completions (
-                    completion_id, session_link_id, child_turn_id, child_last_event_seq,
-                    outcome, created_at, updated_at
-                 ) VALUES (
-                    'completion-1', 'link-1', 'turn-child-1', 42,
-                    'completed', ?1, ?1
-                 )",
-                ["2026-03-25T00:01:06Z"],
-            )?;
-            conn.execute(
-                "INSERT INTO session_link_wake_schedules (session_link_id)
-                 VALUES ('link-1')",
-                [],
-            )?;
-            Ok(())
-        })
-        .expect("insert subagent link and completion");
-
-        assert_eq!(count_rows(&db, "session_events", "session-1"), 1);
-        assert_eq!(count_rows(&db, "session_raw_notifications", "session-1"), 1);
-        assert_eq!(
-            count_rows(&db, "session_live_config_snapshots", "session-1"),
-            1
-        );
-        assert_eq!(
-            count_rows(&db, "session_pending_config_changes", "session-1"),
-            1
-        );
-        assert_eq!(count_rows(&db, "session_pending_prompts", "session-1"), 1);
-        assert_eq!(count_rows(&db, "session_background_work", "session-1"), 1);
-        assert_eq!(count_all_rows(&db, "session_links"), 1);
-        assert_eq!(count_all_rows(&db, "session_link_completions"), 1);
-        assert_eq!(count_all_rows(&db, "session_link_wake_schedules"), 1);
-
-        store
-            .delete_session("session-1")
-            .expect("delete session with dependents");
-
-        assert!(store
-            .find_by_id("session-1")
-            .expect("load deleted session")
-            .is_none());
-        assert_eq!(count_rows(&db, "session_events", "session-1"), 0);
-        assert_eq!(count_rows(&db, "session_raw_notifications", "session-1"), 0);
-        assert_eq!(
-            count_rows(&db, "session_live_config_snapshots", "session-1"),
-            0
-        );
-        assert_eq!(
-            count_rows(&db, "session_pending_config_changes", "session-1"),
-            0
-        );
-        assert_eq!(count_rows(&db, "session_pending_prompts", "session-1"), 0);
-        assert_eq!(count_rows(&db, "session_background_work", "session-1"), 0);
-        assert_eq!(count_all_rows(&db, "session_links"), 0);
-        assert_eq!(count_all_rows(&db, "session_link_completions"), 0);
-        assert_eq!(count_all_rows(&db, "session_link_wake_schedules"), 0);
-    }
-
-    #[test]
-    fn update_title_persists_session_title() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        let mut record = session_record();
-        record.title = None;
-
-        store.insert(&record).expect("insert session");
-        store
-            .update_title(
-                "session-1",
-                "Investigate flaky checkout",
-                "2026-03-25T01:00:00Z",
-            )
-            .expect("update title");
-
-        let stored = store
-            .find_by_id("session-1")
-            .expect("find session")
-            .expect("session record");
-
-        assert_eq!(stored.title.as_deref(), Some("Investigate flaky checkout"));
-        assert_eq!(stored.updated_at, "2026-03-25T01:00:00Z");
-    }
-
-    #[test]
-    fn raw_notifications_are_persisted_in_seq_order() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        let record = session_record();
-        store.insert(&record).expect("insert session");
-
-        store
-            .append_raw_notification(
-                "session-1",
-                "agent_message_chunk",
-                "2026-03-25T00:00:01Z",
-                r#"{"update":{"sessionUpdate":"agent_message_chunk"}}"#,
-            )
-            .expect("append first raw notification");
-        store
-            .append_raw_notification(
-                "session-1",
-                "tool_call",
-                "2026-03-25T00:00:02Z",
-                r#"{"update":{"sessionUpdate":"tool_call"}}"#,
-            )
-            .expect("append second raw notification");
-
-        let all = store
-            .list_raw_notifications("session-1")
-            .expect("list raw notifications");
-        assert_eq!(all.len(), 2);
-        assert_eq!(all[0].seq, 1);
-        assert_eq!(all[0].notification_kind, "agent_message_chunk");
-        assert_eq!(all[1].seq, 2);
-        assert_eq!(all[1].notification_kind, "tool_call");
-
-        let tail = store
-            .list_raw_notifications_after("session-1", 1)
-            .expect("list raw notifications after");
-        assert_eq!(tail.len(), 1);
-        assert_eq!(tail[0].seq, 2);
-    }
-
-    #[test]
-    fn visible_session_lists_exclude_dismissed_and_closed_sessions() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-
-        let visible = session_record();
-        store.insert(&visible).expect("insert visible session");
-
-        let mut dismissed = session_record();
-        dismissed.id = "session-2".to_string();
-        dismissed.dismissed_at = Some("2026-03-25T02:00:00Z".to_string());
-        dismissed.updated_at = "2026-03-25T02:00:00Z".to_string();
-        store.insert(&dismissed).expect("insert dismissed session");
-
-        let mut closed = session_record();
-        closed.id = "session-3".to_string();
-        closed.status = "closed".to_string();
-        closed.closed_at = Some("2026-03-25T03:00:00Z".to_string());
-        closed.updated_at = "2026-03-25T03:00:00Z".to_string();
-        store.insert(&closed).expect("insert closed session");
-
-        let visible_by_workspace = store
-            .list_visible_by_workspace("workspace-1")
-            .expect("list visible sessions by workspace");
-        assert_eq!(visible_by_workspace.len(), 1);
-        assert_eq!(visible_by_workspace[0].id, "session-1");
-
-        let with_dismissed = store
-            .list_with_dismissed_by_workspace("workspace-1")
-            .expect("list sessions with dismissed by workspace");
-        assert_eq!(with_dismissed.len(), 2);
-        assert_eq!(with_dismissed[0].id, "session-2");
-        assert_eq!(with_dismissed[1].id, "session-1");
-    }
-
-    #[test]
-    fn mark_dismissed_is_idempotent_and_restore_uses_latest_timestamp() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-
-        let first = session_record();
-        store.insert(&first).expect("insert first session");
-
-        let mut second = session_record();
-        second.id = "session-2".to_string();
-        store.insert(&second).expect("insert second session");
-
-        store
-            .mark_dismissed("session-1", "2026-03-25T01:00:00Z")
-            .expect("dismiss first session");
-        store
-            .mark_dismissed("session-1", "2026-03-25T05:00:00Z")
-            .expect("repeat dismiss first session");
-        store
-            .mark_dismissed("session-2", "2026-03-25T03:00:00Z")
-            .expect("dismiss second session");
-
-        let first_stored = store
-            .find_by_id("session-1")
-            .expect("find first session")
-            .expect("first session exists");
-        assert_eq!(
-            first_stored.dismissed_at.as_deref(),
-            Some("2026-03-25T01:00:00Z")
-        );
-        assert_eq!(first_stored.updated_at, "2026-03-25T01:00:00Z");
-
-        let last_dismissed = store
-            .find_last_dismissed_in_workspace("workspace-1")
-            .expect("find last dismissed session")
-            .expect("dismissed session exists");
-        assert_eq!(last_dismissed.id, "session-2");
-
-        store
-            .clear_dismissed("session-2", "2026-03-25T04:00:00Z")
-            .expect("restore second session");
-
-        let restored = store
-            .find_by_id("session-2")
-            .expect("find restored session")
-            .expect("restored session exists");
-        assert_eq!(restored.dismissed_at, None);
-        assert_eq!(restored.updated_at, "2026-03-25T04:00:00Z");
-
-        let remaining = store
-            .find_last_dismissed_in_workspace("workspace-1")
-            .expect("find remaining dismissed session")
-            .expect("remaining dismissed session exists");
-        assert_eq!(remaining.id, "session-1");
-    }
-
-    #[test]
-    fn pop_last_dismissed_restores_latest_session_atomically() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-
-        let mut first = session_record();
-        first.id = "session-1".to_string();
-        store.insert(&first).expect("insert first session");
-
-        let mut second = session_record();
-        second.id = "session-2".to_string();
-        store.insert(&second).expect("insert second session");
-
-        store
-            .mark_dismissed("session-1", "2026-03-25T01:00:00Z")
-            .expect("dismiss first session");
-        store
-            .mark_dismissed("session-2", "2026-03-25T03:00:00Z")
-            .expect("dismiss second session");
-
-        let restored = store
-            .pop_last_dismissed_in_workspace("workspace-1", "2026-03-25T04:00:00Z")
-            .expect("pop dismissed session")
-            .expect("restored session exists");
-        assert_eq!(restored.id, "session-2");
-        assert_eq!(restored.dismissed_at, None);
-        assert_eq!(restored.updated_at, "2026-03-25T04:00:00Z");
-
-        let next = store
-            .pop_last_dismissed_in_workspace("workspace-1", "2026-03-25T05:00:00Z")
-            .expect("pop next dismissed session")
-            .expect("next restored session exists");
-        assert_eq!(next.id, "session-1");
-        assert_eq!(next.dismissed_at, None);
-
-        let none = store
-            .pop_last_dismissed_in_workspace("workspace-1", "2026-03-25T06:00:00Z")
-            .expect("pop empty dismissed stack");
-        assert!(none.is_none());
-    }
-
-    #[test]
-    fn background_work_round_trips_and_marks_terminal() {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-
-        let store = SessionStore::new(db);
-        store.insert(&session_record()).expect("insert session");
-
-        let pending = SessionBackgroundWorkRecord {
-            session_id: "session-1".to_string(),
-            tool_call_id: "tool-1".to_string(),
-            turn_id: "turn-1".to_string(),
-            tracker_kind: SessionBackgroundWorkTrackerKind::ClaudeAsyncAgent,
-            source_agent_kind: "claude".to_string(),
-            agent_id: Some("agent-1".to_string()),
-            output_file: "/tmp/agent.output".to_string(),
-            state: SessionBackgroundWorkState::Pending,
-            created_at: "2026-03-25T01:00:00Z".to_string(),
-            updated_at: "2026-03-25T01:00:00Z".to_string(),
-            launched_at: "2026-03-25T01:00:00Z".to_string(),
-            last_activity_at: "2026-03-25T01:00:00Z".to_string(),
-            completed_at: None,
-        };
-
-        assert!(store
-            .upsert_or_refresh_pending_background_work(&pending)
-            .expect("upsert pending background work"));
-        store
-            .touch_background_work_activity("session-1", "tool-1", "2026-03-25T01:05:00Z")
-            .expect("touch background work activity");
-
-        let pending_rows = store
-            .list_pending_background_work("session-1")
-            .expect("list pending background work");
-        assert_eq!(pending_rows.len(), 1);
-        assert_eq!(pending_rows[0].last_activity_at, "2026-03-25T01:05:00Z");
-        assert_eq!(pending_rows[0].updated_at, "2026-03-25T01:05:00Z");
-
-        assert!(store
-            .mark_background_work_terminal(
-                "session-1",
-                "tool-1",
-                SessionBackgroundWorkState::Completed,
-                "2026-03-25T01:06:00Z",
-            )
-            .expect("mark background work terminal"));
-
-        assert!(store
-            .list_pending_background_work("session-1")
-            .expect("list pending background work")
-            .is_empty());
-    }
-}
