+++ scripts/max_lines_allowlist.txt
@@ -23,7 +23,6 @@ anyharness/crates/anyharness-lib/src/sessions/live_config.rs
 anyharness/crates/anyharness-lib/src/sessions/prompt.rs
 anyharness/crates/anyharness-lib/src/sessions/runtime.rs
 anyharness/crates/anyharness-lib/src/sessions/service.rs
-anyharness/crates/anyharness-lib/src/sessions/store.rs
 anyharness/crates/anyharness-lib/src/terminals/service.rs
 anyharness/crates/anyharness-lib/src/domains/cowork/runtime.rs
 anyharness/crates/anyharness-lib/src/domains/mobility/service.rs
